requirements.txt
pandas
numpy
scikit-learn
xgboost
optuna
tensorflow
nltk
newsapi-python
MetaTrader5
python-dotenv
requests
joblib
schedule
python-telegram-bot
stable-baselines3
gym
feedparser
